TimeUpdated_MovementUpdate
-Movement Smoother with Delta Time
-Refined the Time Script with Chrono and implemented deltatime

+++ WhyKnotTest_Chiew/Board.cpp
@@ -59,7 +59,7 @@ Board::Board(int widthSize, int heightSize)
 	FoodList.push_back(food3);
 }
 
-void Board::Update(int elapsedTime) //update all the snake and food and perform draw board with the update
+void Board::Update(float elapsedTime) //update all the snake and food and perform draw board with the update
 {
 	snake->Update(elapsedTime);
 	drawBoard();
